// src/common/enums/alerta.enum.ts
export enum TipoAlerta {
  STOCK_BAJO = 'STOCK_BAJO',
  PRODUCTO_AGOTADO = 'PRODUCTO_AGOTADO',
  CUOTA_POR_VENCER = 'CUOTA_POR_VENCER',
  CUOTA_VENCIDA = 'CUOTA_VENCIDA',
  CLIENTE_LIMITE_CREDITO = 'CLIENTE_LIMITE_CREDITO',
  VENTA_EN_MORA = 'VENTA_EN_MORA',
  META_VENTAS_NO_ALCANZADA = 'META_VENTAS_NO_ALCANZADA',
  PERSONALIZADA = 'PERSONALIZADA',
  REGLA = 'REGLA',
}

export enum SeveridadAlerta {
  BAJA = 'BAJA',
  MEDIA = 'MEDIA',
  ALTA = 'ALTA',
  CRITICA = 'CRITICA',
}